// client/src/state/constants/userConstants.ts
export enum UserLoginConstant {
  USER_LOGIN_REQUEST = "USER_LOGIN_REQUEST",
  USER_LOGIN_SUCCESS = "USER_LOGIN_SUCCESS",
  USER_LOGIN_FAIL = "USER_LOGIN_FAIL",
  USER_LOGOUT = "USER_LOGOUT",
  USER_LOGOUT_FAIL = "USER_LOGOUT_FAIL",
}

export enum UserRegisterContsant {
  USER_REGISTER_REQUEST = "USER_REGISTER_REQUEST",
  USER_REGISTER_SUCCESS = "USER_REGISTER_SUCCESS",
  USER_REGISTER_FAIL = "USER_REGISTER_FAIL",
}

export enum UserProfileConstant {
  USER_PROFILE_REQUEST = "USER_PROFILE_REQUEST",
  USER_PROFILE_SUCCESS = "USER_PROFILE_SUCCESS",
  USER_PROFILE_FAIL = "USER_PROFILE_FAIL",
  USER_PROFILE_RESET = "USER_PROFILE_RESET",
}

export enum ProfileUpdateConstant {
  USER_PROFILE_UPADTE_REQUEST = "USER_PROFILE_UPADTE_REQUEST",
  USER_PROFILE_UPADTE_SUCCESS = "USER_PROFILE_UPADTE_SUCCESS",
  USER_PROFILE_UPADTE_FAIL = "USER_PROFILE_UPADTE_FAIL",
  USER_PROFILE_UPADTE_RESET = "USER_PROFILE_UPADTE_RESET",
}

export enum UserForgetPassword {
  FORGOT_PASSWORD_REQUEST = "FORGOT_PASSWORD_REQUEST",
  FORGOT_PASSWORD_SUCCESS = "FORGOT_PASSWORD_SUCCESS",
  FORGOT_PASSWORD_FAIL = "FORGOT_PASSWORD_FAIL",
}

export enum UserResetPassword {
  RESET_PASSWORD_REQUEST = "RESET_PASSWORD_REQUEST",
  RESET_PASSWORD_SUCCESS = "RESET_PASSWORD_SUCCESS",
  RESET_PASSWORD_FAIL = "RESET_PASSWORD_FAIL",
}